Bug fix: multiple workflow on same page causes errors

+++ web/concrete/src/Workflow/BasicWorkflow.php
@@ -64,28 +64,34 @@ class BasicWorkflow extends \Concrete\Core\Workflow\Workflow
     {
         // lets save the basic data associated with this workflow.
         $req = $wp->getWorkflowRequestObject();
-        $db = Loader::db();
-        $db->Execute(
-            'INSERT INTO BasicWorkflowProgressData (wpID, uIDStarted) VALUES (?, ?)',
-            array($wp->getWorkflowProgressID(), $req->getRequesterUserID()));
+        
+        // Check if the workflow is not already approved
+        if (is_object($req)) {
+        
+            $db = Loader::db();
+            $db->Execute(
+                'INSERT INTO BasicWorkflowProgressData (wpID, uIDStarted) VALUES (?, ?)',
+                array($wp->getWorkflowProgressID(), $req->getRequesterUserID()));
 
-        if ($this->canApproveWorkflow()) {
-            // Then that means we have the ability to approve the workflow we just started.
-            // In that case, we transparently approve it, and skip the entry notification step.
-            $wpr = $req->approve($wp);
-            $wp->delete();
+            if ($this->canApproveWorkflow()) {
+                // Then that means we have the ability to approve the workflow we just started.
+                // In that case, we transparently approve it, and skip the entry notification step.
+                $wpr = $req->approve($wp);
+                $wp->delete();
 
-        } else {
+            } else {
 
-            $ui = UserInfo::getByID($req->getRequesterUserID());
+                $ui = UserInfo::getByID($req->getRequesterUserID());
 
-            // let's get all the people who are set to be notified on entry
-            $message = t(
-                'On %s, user %s submitted the following request: %s',
-                Core::make('helper/date')->formatDateTime($wp->getWorkflowProgressDateAdded(), true),
-                $ui->getUserName(),
-                $req->getWorkflowRequestDescriptionObject()->getEmailDescription());
-            $this->notify($wp, $message, 'notify_on_basic_workflow_entry');
+                // let's get all the people who are set to be notified on entry
+                $message = t(
+                    'On %s, user %s submitted the following request: %s',
+                    Core::make('helper/date')->formatDateTime($wp->getWorkflowProgressDateAdded(), true),
+                    $ui->getUserName(),
+                    $req->getWorkflowRequestDescriptionObject()->getEmailDescription());
+                $this->notify($wp, $message, 'notify_on_basic_workflow_entry');
+            }
+        
         }
 
     }
